main.c
The Stopwatch Project is an implementation of a stopwatch using an AVR microcontroller. The project utilizes external interrupts and Timer1 to track time with second precision and displays the elapsed time on a 7-segment display.

Usage:

To start the stopwatch, press the start button.
To stop the stopwatch, press the stop button.
To reset the stopwatch, press the reset button.